reorganzie _report_utils
