+++ src/vampire/_report_utils/_make_report.py
@@ -8,10 +8,12 @@ def make_report(job_dir: str, template_html_path: str, data: dict) -> None:
     Output:
         None
     """
-
     with open(template_html_path, 'r') as f:
         template = f.read()
     for key, value in data.items():
         template = template.replace(f"${key}$", str(value))
     with open(f"{job_dir}/web_summary.html", 'w') as f:
         f.write(template)
+
+
+
