+++ src/vampire/_report_utils/_stats_seq.py
