+++ src/vampire/_report_utils/_table.py
